requirements.txt
-i https://pypi.python.org/simple
annotated-types==0.7.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
boto3==1.36.26; python_version >= '3.8'
botocore==1.36.26; python_version >= '3.8'
diff-match-patch==20241021; python_version >= '3.7'
dj-database-url==2.3.0
django==5.1.6; python_version >= '3.10'
django-appconf==1.1.0; python_version >= '3.9'
django-axes==7.0.2; python_version >= '3.7'
django-cors-headers==4.7.0; python_version >= '3.9'
django-dbsettings==1.3.0
django-filter==25.1; python_version >= '3.9'
django-imagekit==5.0.0
django-import-export==4.3.6; python_version >= '3.9'
django-qr-code==4.1.0; python_version >= '3.10'
django-storages==1.14.5; python_version >= '3.7'
django-taggit==6.1.0; python_version >= '3.8'
django-widget-tweaks==1.5.0; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
faker==36.1.1; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
pilkit==3.0
pillow==11.1.0; python_version >= '3.9'
plotly==5.24.1; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
s3transfer==0.11.2; python_version >= '3.8'
segno==1.6.1; python_version >= '3.5'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.3; python_version >= '3.8'
tablib==3.8.0; python_version >= '3.9'
tenacity==9.0.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.10'
paho-mqtt>=2.1.0; python_version >= '3.8'
whitenoise==6.9.0; python_version >= '3.9'

# MCP dev server (optional; install with: pip install fastmcp)
# fastmcp
